requirements.txt
certifi==2019.11.28
chardet==3.0.4
docopt==0.6.2
elasticsearch==6.8.1
elasticsearch-dsl==6.4.0
idna==2.9
psycopg2==2.8.5
pymarc==4.0.0
python-dateutil==2.8.1
python-dotenv==0.13.0
redis==3.5.2
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.17
stomp.py==6.0.0
tqdm==4.46.0
ujson==2.0.3
urllib3==1.25.9
